+++ sentry_sdk/consts.py
@@ -104,7 +104,7 @@ DEFAULT_OPTIONS = _get_default_options()
 del _get_default_options
 
 
-VERSION = "1.10.0"
+VERSION = "1.10.1"
 
 
 class OP: